fix crash error : wrong pointer operation!

+++ musicplayer/src/main/cpp/RayQueue.h
@@ -27,7 +27,8 @@ public:
     ~RayQueue();
     void putPacket(AVPacket * packet);
     int getPacket(AVPacket * packet);
-    int getSize();
+
+    long getSize();
 
 };
